Optimise pipeline

+++ src/Annotater.py
@@ -19,28 +19,50 @@ class Annotater:
         self.configDir = configDir
         self.run_type = run_type
         self.local = local
-        self.fileName = os.path.basename(mutTarget)
-        self.parentDirectoryPath = os.path.dirname(mutTarget)
-        self.hgvsFilePath = mutTarget + '.hgvs'
-        self.vcfFilePath = mutTarget + '.vcf'
-        self.ensemblFilePath = mutTarget + '.ensembl'
+        vcf_cols = ["#chrom", "pos", "id", "ref", "alt", "qual", "filter", "info"]
+        ensembl_cols = ["#chrom", "pos", "end", "ref/alt", "strand", "id"]
+        if os.path.isfile(self.mutTarget):
+            target = os.path.dirname(self.mutTarget)
+        else:
+            target = mutTarget
+        self.hgvsFilePath = os.path.join(target, 'merged.hgvs')
+        self.vcfFilePath = os.path.join(target, 'merged.vcf')
+        self.ensemblFilePath = os.path.join(target, 'merged.ensembl')
+        self.annFilePath = os.path.join(target, 'merged.ANN')
+        self.vcfDf = pd.DataFrame(columns=vcf_cols)
+        self.ensemblDf = pd.DataFrame(columns=ensembl_cols)
 
-    def run(self):
-        logging.basicConfig(filename='{}.log'.format(self.mutTarget), filemode='a+', level=logging.DEBUG)
-        logging.info("Starting merge of annotations")
+    def run(self, mutFile):
+        if mutFile != '':
+            self.mutTarget = join(self.mutTarget, mutFile)
+        self.fileName = os.path.basename(self.mutTarget)
+        self.parentDirectoryPath = os.path.dirname(self.mutTarget)
+        logging.basicConfig(filename='{}.log'.format(join(self.parentDirectoryPath, 'annotater')), filemode='a+', level=logging.DEBUG)
+        logging.info("Starting merge of annotations using " + str(cpu_count())+" CPUs")
         if self.run_type == 'vcf':
             self.process_VCForEnsembl()
         elif self.run_type == 'hgvs':
             self.process_hgvs()
+
+    def annotate(self):
+        if self.run_type == 'hgvs':
+            self.annotateFile(self.hgvsFilePath, 'hgvs')
+            self.process_hgvs_annotations()
+            os.remove(self.hgvsFilePath)
+        elif self.run_type == 'vcf':
+            self.annotateFile(self.vcfFilePath, "vcf")
+            self.annotateFile(self.ensemblFilePath, "ensembl")
+            self.mergeResultAnnos(self.vcfFilePath, self.ensemblFilePath)
         logging.info("Annotating is complete")
+
 
     def process_hgvs(self):
         self.formatHGVSFiles()
         self.dropDuplicates(self.hgvsFilePath)
         #'singularity exec -B /Users/tushar/PycharmProjects/PDCMDataAggregator/output,/Users/tushar/pdx/update-data,/Users/tushar/pdx/mutation-annotater:/Users/tushar/pdx/mutation-annotater:rw pdx-liftover_vep_release98.3.sif vep --verbose --vcf --force --check_existing --check_ref --refseq --use_given_ref --symbol --polyphen s --sift s --hgvs --variant_class --no_stats --pick --pick_order biotype,canonical,appris,tsl,ccds,rank,length,mane --fork=16 --warning_file  /Users/tushar/PycharmProjects/PDCMDataAggregator/output/CMP/pdcm_format/mut/CMP_mut.tsv.ensembl.vepWarnings --cache --dir_cache /Users/tushar/pdx/mutation-annotater/vepDB/homo_sapiens_vep_98_GRCh38 --fasta /Users/tushar/pdx/mutation-annotater/vepDB/homo_sapiens/Homo_sapiens.GRCh38.dna.primary_assembly.fa --format hgvs -i /Users/tushar/PycharmProjects/PDCMDataAggregator/output/CMP/pdcm_format/mut/WGS/CMP_mut_SIDS01895.tsv.refseq -o /Users/tushar/PycharmProjects/PDCMDataAggregator/output/CMP/pdcm_format/mut/WGS/CMP_mut_SIDS01895.tsv.ANN 2>> /Users/tushar/PycharmProjects/PDCMDataAggregator/output/CMP/pdcm_format/mut/WGS/CMP_mut_SIDS01895.tsv.log'
-        self.annotateFile(self.hgvsFilePath, 'hgvs')
-        self.process_hgvs_annotations()
-        os.remove(self.hgvsFilePath)
+        #self.annotateFile(self.hgvsFilePath, 'hgvs')
+        #self.process_hgvs_annotations()
+        #os.remove(self.hgvsFilePath)
 
     def process_hgvs_annotations(self):
         mergedAnnos = self.mutTarget + ".ANN"
@@ -59,7 +81,6 @@ class Annotater:
             vcfDf.join(infoColumnsDf).rename(columns=str.lower).to_csv(mergedAnnos, sep='\t', index=False)
             os.remove(vcfAnnos)
 
-
     def formatHGVSFiles(self):
         with open(self.hgvsFilePath, 'w') as hgvsFile:
             reader = pd.read_csv(self.mutTarget, sep='\t', dtype=str, engine='python').fillna('')
@@ -72,52 +93,55 @@ class Annotater:
             message = "The file {0} has {1} data points (including header)".format(self.mutTarget, (index + 1))
             logging.info(message)
 
-
     def process_VCForEnsembl(self):
         self.formatToVcfOrEnsemblAndSave()
-        self.processFiles()
-        self.annotateFile(self.vcfFilePath, "vcf")
-        self.annotateFile(self.ensemblFilePath, "ensembl")
-        self.mergeResultAnnos(self.vcfFilePath, self.ensemblFilePath)
+        #self.processFiles()
+        #self.annotateFile(self.vcfFilePath, "vcf")
+        #self.annotateFile(self.ensemblFilePath, "ensembl")
+        #self.mergeResultAnnos(self.vcfFilePath, self.ensemblFilePath)
 
     def formatToVcfOrEnsemblAndSave(self):
-        with open(self.vcfFilePath, "w") as vcfFile, \
-                open(self.ensemblFilePath, "w") as ensembl:
-            reader = pd.read_csv(self.mutTarget, sep='\t', dtype=str, engine='python')
-            reader['chromosome'] = reader.apply(lambda x: self.formatChromo(x.chromosome), axis=1)
-            logging.info("Writing {0} to VCF".format(self.mutTarget))
-            vcfFile.write("#chrom\tpos\tid\tref\talt\tqual\tfilter\tinfo\n")
-            ensembl.write("#chrom\tpos\tend\tref/alt\tstrand\tid\n")
-            for index, row in reader.iterrows():
-                rowDict = row.to_dict()
-                if self.isRowValidForProcessing(rowDict):
-                    if self.rowIsEnsembl(rowDict):
-                        self.formatRowToEnsemblAndWrite(rowDict, ensembl)
-                    else:
-                        self.formatRowToVCFAndWrite(rowDict, vcfFile)
-            message = "The file {0} has {1} data points (including header)".format(self.mutTarget, (index + 1))
-            logging.info(message)
+        reader = pd.read_csv(self.mutTarget, sep='\t', dtype=str, engine='python')
+        reader = self.isRowValidForProcessing(reader)
+        reader['chromosome'] = reader.apply(lambda x: self.formatChromo(x.chromosome), axis=1)
+        reader['id'] = reader.apply(
+            lambda x: "{}_{}_{}_{}".format(self.formatChromo(x["chromosome"]), x["seq_start_position"],
+                                           x["ref_allele"], x["alt_allele"]), axis=1)
 
-    def rowIsEnsembl(self, row):
-        refAllele = row["ref_allele"]
-        altAllele = row["alt_allele"]
-        return bool(re.search('-', refAllele)) or bool(re.search('-', altAllele))
+        logging.info("Writing {0} to VCF".format(self.mutTarget))
+        self.generate_ensembl_file(reader)
+        self.generate_vcf_file(reader)
 
-    def isRowValidForProcessing(self, row):
-        isValid = True
-        if self.anyGenomicCoordinateAreMissing(row):
+        #for index, row in reader.iterrows():
+        #    rowDict = row.to_dict()
+        #    if self.isRowValidForProcessing(rowDict):
+        #        if self.rowIsEnsembl(rowDict):
+        #            self.formatRowToEnsemblAndWrite(rowDict, ensembl)
+        #        else:
+        #            self.formatRowToVCFAndWrite(rowDict, vcfFile)
+
+        message = "The file {0} has {1} data points (including header)".format(self.mutTarget, (reader.shape[0] + 1))
+        logging.info(message)
+
+    def isRowValidForProcessing(self, df):
+        indices_before = df.index
+        df = df.dropna(subset=['chromosome', 'seq_start_position', 'ref_allele', 'alt_allele'])
+        indices_after = df.index
+        dropped_indices = indices_before.difference(indices_after)
+        for index in dropped_indices:
             logging.info(
-                "Row has incomplete data : {0} in file {1} caused by missing chro,seq start, ref or alt allele data"
-                .format(row.items(), self.vcfFilePath))
-            isValid = False
-        return isValid
+                "Row has incomplete data : Row {0} in file {1} has missing chr, seq start, ref or alt allele data"
+                .format(index, self.mutTarget))
+        return df
 
     def processFiles(self):
         logging.info("removing duplicates in VCF/Ensembl files")
-        self.sortInPlace(self.vcfFilePath)
-        self.sortInPlace(self.ensemblFilePath)
-        self.dropDuplicates(self.vcfFilePath)
-        self.dropDuplicates(self.ensemblFilePath)
+        self.vcfDf = sort_vcf_ensembl_df(self.vcfDf)
+        self.ensemblDf = sort_vcf_ensembl_df(self.ensemblDf)
+        self.vcfDf.to_csv(self.vcfFilePath, sep='\t', index=False)
+        self.ensemblDf.to_csv(self.ensemblFilePath, sep='\t', index=False)
+        #self.sortInPlace(self.vcfFilePath)
+        #self.sortInPlace(self.ensemblFilePath)
 
     def sortInPlace(self, aVCFfile):
         with open(aVCFfile, 'r') as infile:
@@ -153,6 +177,34 @@ class Annotater:
                                                              self.createPosId(row),
                                                              row["ref_allele"], row["alt_allele"])
         vcfFile.write(vcfRow)
+
+    def generate_vcf_file(self, df):
+        temp = df[~df['ref_allele'].str.contains('-') & ~df['alt_allele'].str.contains('-')]
+        if len(temp) > 0:
+            temp = temp[['chromosome', 'strand', 'seq_start_position', 'ref_allele', 'alt_allele', 'id']]
+            mapper = {'chromosome': '#chrom', 'seq_start_position': 'pos',
+                      'ref_allele': 'ref', 'alt_allele': 'alt'}
+            temp.rename(columns=mapper, inplace=True)
+            temp['qual'] = ''
+            temp['filter'] = ''
+            temp['info'] = ''
+            #temp['strand'] = ['1' if r == "" else str(int(r)) for r in temp['strand']]
+            cols = ["#chrom", "pos", "id", "ref", "alt", "qual", "filter", "info"]
+            self.vcfDf = pd.concat([self.vcfDf, temp[cols]])
+            self.vcfDf = self.vcfDf.drop_duplicates().reset_index(drop=True)
+
+    def generate_ensembl_file(self, df):
+        temp = df[df['ref_allele'].str.contains('-') | df['alt_allele'].str.contains('-')]
+        if len(temp)>0:
+            temp = temp[['chromosome', 'strand', 'seq_start_position', 'ref_allele', 'alt_allele', 'id']]
+            temp['strand'] = ['1' if r == "" else str(int(r)) for r in temp['strand'].fillna('')]
+            temp['pos'] = temp['seq_start_position']
+            temp['end'] = temp.apply(lambda x: int(x['pos']) - 1 if bool(re.search("-", x["ref_allele"])) else int(x['pos']) + len(x["ref_allele"]) - 1 if bool(re.search("-", x["alt_allele"])) else "" , axis=1)
+            temp['ref/alt'] = temp['ref_allele'].astype(str) + '/' + temp['alt_allele'].astype(str)
+            temp['#chrom'] = temp['chromosome']
+            cols = ['#chrom', 'pos', 'end', 'ref/alt', 'strand', 'id']
+            self.ensemblDf = pd.concat([self.ensemblDf, temp[cols]])
+            self.ensemblDf = self.ensemblDf.drop_duplicates().reset_index(drop=True)
 
     def formatRowToEnsemblAndWrite(self, row, ensemblFile):
         startPos, endPos = self.resolveEnsemblEndPos(row)
@@ -200,7 +252,7 @@ class Annotater:
         vepOut = vepIn + ".ANN"
         threads = cpu_count() * 2
         vepCMD = """vep {0} --format {1} --fork={2} --warning_file {3} --cache --dir_cache {4} --fasta {5} -i {6} -o {7} 2>> {8}.log""" \
-            .format(vepArguments, format, threads, vepWarningFile, alleleDB, fastaDir, vepIn, vepOut, self.mutTarget)
+            .format(vepArguments, format, threads, vepWarningFile, alleleDB, fastaDir, vepIn, vepOut, join(self.parentDirectoryPath, 'annotater'))
         if format != 'hgvs':
             vepCMD = vepCMD + " --offline --merged " #Get Refseq and Ensembl lookup databases
         elif format == 'hgvs':
@@ -250,7 +302,7 @@ class Annotater:
     def mergeResultAnnos(self, vcfPath, ensemblPath):
         vcfAnnos = vcfPath + ".ANN"
         ensemblAnnos = ensemblPath + ".ANN"
-        mergedAnnos = self.mutTarget + ".ANN"
+        mergedAnnos = self.annFilePath
         with open(vcfAnnos, 'r') as vcfFile:
             vcfFile.readline()
             vcfFile.readline()
@@ -271,18 +323,30 @@ def cmdline_runner():
         run_type = sys.argv[2]
         local = sys.argv[3]
         local = local == "local"
+        annotate = Annotater(mutTarget, run_type, local)
         if os.path.isfile(mutTarget):
             logging.basicConfig(filename='{}.log'.format(mutTarget), filemode='a+', level=logging.DEBUG)
             logging.info(" Starting annotation pipleline ")
-            Annotater(mutTarget, run_type, local).run()
+            annotate.run('')
+            annotate.processFiles()
+            annotate.annotate()
         elif os.path.isdir(mutTarget):
             globForTsv = os.path.join(mutTarget, "*tsv")
             for mutFile in glob.iglob(globForTsv):
-                mutfile_path = os.path.join(mutTarget, mutFile)
-                print(mutfile_path)
-                Annotater(mutfile_path, run_type, local,'config.yaml').run()
+                #mutfile_path = os.path.join(mutTarget, mutFile)
+                #print(mutfile_path)
+                annotate.run(mutFile)
+            annotate.processFiles()
+            annotate.annotate()
+
     else:
         logging.info("Please pass the absolute path of the file to annotate")
 
+def sort_vcf_ensembl_df(df):
+    cols = df.columns
+    df['chromosome'] = pd.Categorical(df['#chrom'], ordered=True,
+                                              categories=['chr' + str(i) for i in range(1, 23)] + ['chrX', 'chrY'])
+    df = df.sort_values(by=['chromosome', 'pos'])
+    return df[cols]
 
 cmdline_runner()
